Day 17/requirements.txt
numpy
matplotlib
scipy
pillow
opencv-python
streamlit
